+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/dsr/repository/user_login/auth/AuthRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthRepository extends JpaRepository<Auth, Long> {
+public interface AuthRepository extends JpaRepository<Auth, Integer> {
 
 }
