+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/dsr/repository/user_login/time/TimeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TimeRepository extends JpaRepository<Time, Long> {
+public interface TimeRepository extends JpaRepository<Time, Integer> {
 }
